decoupling more engines

+++ plugins/adva_google_analytics/db/migrate/20081215155500_add_google_analytics_to_issues.rb
@@ -1,14 +1,17 @@
 class AddGoogleAnalyticsToIssues < ActiveRecord::Migration
   def self.up
-    # FIXME check if table exists!
-    add_column :issues, :track, :boolean, :default => false
-    add_column :issues, :tracking_campaign, :string
-    add_column :issues, :tracking_source, :string
+    if connection.table_exists?(:issues)
+      add_column :issues, :track, :boolean, :default => false
+      add_column :issues, :tracking_campaign, :string
+      add_column :issues, :tracking_source, :string
+    end
   end
 
   def self.down
-    remove_column :issues, :track
-    remove_column :issues, :tracking_campaign
-    remove_column :issues, :tracking_source
+    if connection.table_exists?(:issues)
+      remove_column :issues, :track
+      remove_column :issues, :tracking_campaign
+      remove_column :issues, :tracking_source
+    end
   end
 end
